Refactor: url interactions post-put
