+++ tclothes/clothes/views/clothes.py
@@ -1,7 +1,10 @@
 """Clothes views."""
 
 # Django REST Framework
-from rest_framework import viewsets, mixins
+from rest_framework import viewsets, mixins, status
+from rest_framework.response import Response
+from rest_framework.decorators import action
+from tclothes.clothes.serializers import InteractionsModelSerializer
 
 # Permissions
 from rest_framework.permissions import IsAuthenticated
@@ -34,3 +37,26 @@ class ClothesViewSet(mixins.ListModelMixin,
     ordering_fields = ['likes', 'size', 'gender', 'created_at']
     ordering = ['-likes']
     filter_fields = ['brand']
+
+    @action(detail=False, methods=['post', 'put'])
+    def interactions(self, request, *args, **kwargs):
+        user = request.user
+        serializer = InteractionsModelSerializer(
+            data=request.data,
+            partial=False
+        )
+        serializer.is_valid(raise_exception=True)
+        serializer.save(user=user)
+
+        # Update stats clothe
+        clothe = ClothesModel.objects.get(id=request.data['clothe'])
+        user_action = request.data['value']
+        if user_action == 'LIKE':
+            clothe.likes += 1
+        elif user_action == 'SUPERLIKE':
+            clothe.likes += 10
+        else:
+            clothe.dislikes += 1
+        clothe.save()
+
+        return Response(status=status.HTTP_202_ACCEPTED)
