+++ tclothes/clothes/views/users_clothes.py
@@ -3,7 +3,6 @@
 # Django REST Framework
 from rest_framework import viewsets, mixins, status
 from rest_framework.response import Response
-from rest_framework.decorators import action
 
 # Permissions
 from rest_framework.permissions import IsAuthenticated
@@ -13,7 +12,7 @@ from rest_framework.permissions import IsAuthenticated
 from tclothes.clothes.models import ClothesModel
 
 # Serializer
-from tclothes.clothes.serializers import ClotheModelSerializer, InteractionsModelSerializer
+from tclothes.clothes.serializers import ClotheModelSerializer
 
 
 class UsersClothesViewSet(mixins.CreateModelMixin,
@@ -41,26 +40,3 @@ class UsersClothesViewSet(mixins.CreateModelMixin,
         clothe = serializer.save(owner_is=request.user)
         data = self.get_serializer(clothe).data
         return Response(data, status=status.HTTP_201_CREATED)
-
-    @action(detail=False, methods=['post', 'put'])
-    def interactions(self, request, *args, **kwargs):
-        user = request.user
-        serializer = InteractionsModelSerializer(
-            data=request.data,
-            partial=False
-        )
-        serializer.is_valid(raise_exception=True)
-        serializer.save(user=user)
-
-        # Update stats clothe
-        clothe = ClothesModel.objects.get(id=request.data['clothe'])
-        user_action = request.data['value']
-        if user_action == 'LIKE':
-            clothe.likes += 1
-        elif user_action == 'SUPERLIKE':
-            clothe.likes += 10
-        else:
-            clothe.dislikes += 1
-        clothe.save()
-
-        return Response(status=status.HTTP_202_ACCEPTED)
